requirements.txt
# Core dependencies
langchain>=0.3.15,<0.4.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.1,<0.3.0
langchain-huggingface>=0.1.2

# Document processing
unstructured[pdf-full]>=0.10.25
pdf2image>=1.16.3
pdfminer.six==20240706
pytesseract>=0.3.10

# Vector store and embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM related
openai>=1.58.1,<2.0.0
tiktoken>=0.7,<1

# Development
python-dotenv>=1.0.0
pytest>=7.4.0

# OCR and PDF processing
pytesseract>=0.3.10

# NLP resources
nltk>=3.6.0